refactor(CH): move types to a separate module
Plus, add implementations for Direction and Position
