+++ frame/clearing-house/src/lib.rs
@@ -114,6 +114,8 @@ mod tests;
 
 mod math;
 
+mod types;
+
 mod weights;
 
 #[cfg(feature = "runtime-benchmarks")]
@@ -125,6 +127,7 @@ pub mod pallet {
 	//                                       Imports and Dependencies
 	// ----------------------------------------------------------------------------------------------------
 
+	pub use crate::types::{Direction, Market, MarketConfig, Position};
 	use crate::{
 		math::{FromBalance, IntoBalance, IntoDecimal, IntoSigned, TryMath},
 		weights::WeightInfo,
@@ -134,7 +137,6 @@ pub mod pallet {
 		clearing_house::{ClearingHouse, Instruments},
 		defi::DeFiComposableConfig,
 		oracle::Oracle,
-		time::DurationSeconds,
 		vamm::{AssetType, Direction as VammDirection, SwapConfig, SwapSimulationConfig, Vamm},
 	};
 	use frame_support::{
@@ -241,89 +243,6 @@ pub mod pallet {
 	// ----------------------------------------------------------------------------------------------------
 	//                                           Pallet Types
 	// ----------------------------------------------------------------------------------------------------
-
-	#[derive(Encode, Decode, TypeInfo, Debug, Clone, Copy, PartialEq)]
-	pub enum Direction {
-		Long,
-		Short,
-	}
-
-	/// Stores the user's position in a particular market
-	#[derive(Encode, Decode, MaxEncodedLen, TypeInfo, Debug)]
-	#[scale_info(skip_type_params(T))]
-	#[codec(mel_bound())]
-	pub struct Position<T: Config> {
-		/// The Id of the virtual market
-		pub market_id: T::MarketId,
-		/// Virtual base asset amount. Positive implies long position and negative, short.
-		pub base_asset_amount: T::Decimal,
-		/// Virtual quote asset notional amount (margin * leverage * direction) used to open the
-		/// position
-		pub quote_asset_notional_amount: T::Decimal,
-		/// Last cumulative funding rate used to update this position. The market's latest
-		/// cumulative funding rate minus this gives the funding rate this position must pay. This
-		/// rate multiplied by this position's size (base asset amount * amm price) gives the total
-		/// funding owed, which is deducted from the trader account's margin. This debt is
-		/// accounted for in margin ratio calculations, which may lead to liquidation.
-		pub last_cum_funding: T::Decimal,
-	}
-
-	/// Data relating to a perpetual contracts market
-	#[derive(Encode, Decode, MaxEncodedLen, TypeInfo, Debug)]
-	#[scale_info(skip_type_params(T))]
-	#[codec(mel_bound())]
-	pub struct Market<T: Config> {
-		/// The Id of the vAMM used for price discovery in the virtual market
-		pub vamm_id: T::VammId,
-		/// The Id of the underlying asset (base-quote pair). A price feed from one or more oracles
-		/// must be available for this symbol
-		pub asset_id: AssetIdOf<T>,
-		/// Minimum margin ratio for opening a new position
-		pub margin_ratio_initial: T::Decimal,
-		/// Margin ratio below which liquidations can occur
-		pub margin_ratio_maintenance: T::Decimal,
-		/// Minimum amount of quote asset to exchange when opening a position. Also serves to round
-		/// a trade if it results in closing an existing position
-		pub minimum_trade_size: T::Decimal,
-		/// The latest cumulative funding rate of this market. Must be updated periodically.
-		pub cum_funding_rate: T::Decimal,
-		/// The timestamp for the latest funding rate update.
-		pub funding_rate_ts: DurationSeconds,
-		/// The time span between each funding rate update.
-		pub funding_frequency: DurationSeconds,
-		/// Period of time over which funding (the difference between mark and index prices) gets
-		/// paid.
-		///
-		/// Setting the funding period too long may cause the perpetual to start trading at a
-		/// very dislocated price to the index because there’s less of an incentive for basis
-		/// arbitrageurs to push the prices back in line since they would have to carry the basis
-		/// risk for a longer period of time.
-		///
-		/// Setting the funding period too short may cause nobody to trade the perpetual because
-		/// there’s too punitive of a price to pay in the case the funding rate flips sign.
-		pub funding_period: DurationSeconds,
-	}
-
-	/// Specifications for market creation
-	#[derive(Encode, Decode, PartialEq, Clone, Debug, TypeInfo)]
-	pub struct MarketConfig<AssetId, Decimal, VammConfig> {
-		/// Asset id of the underlying for the derivatives market
-		pub asset: AssetId,
-		/// Configuration for creating and initializing the vAMM for price discovery
-		pub vamm_config: VammConfig,
-		/// Minimum margin ratio for opening a new position
-		pub margin_ratio_initial: Decimal,
-		/// Margin ratio below which liquidations can occur
-		pub margin_ratio_maintenance: Decimal,
-		/// Minimum amount of quote asset to exchange when opening a position. Also serves to round
-		/// a trade if it results in closing an existing position
-		pub minimum_trade_size: Decimal,
-		/// Time span between each funding rate update
-		pub funding_frequency: DurationSeconds,
-		/// Period of time over which funding (the difference between mark and index prices) gets
-		/// paid.
-		pub funding_period: DurationSeconds,
-	}
 
 	type AssetIdOf<T> = <T as DeFiComposableConfig>::MayBeAssetId;
 	type BalanceOf<T> = <T as DeFiComposableConfig>::Balance;
@@ -760,7 +679,7 @@ pub mod pallet {
 				Error::<T>::TradeSizeTooSmall
 			);
 
-			let position_direction = Self::position_direction(position).unwrap_or(direction);
+			let position_direction = position.direction().unwrap_or(direction);
 
 			let base_swapped: T::Balance;
 			// Whether or not the trade increases the risk exposure of the account
@@ -1039,7 +958,7 @@ pub mod pallet {
 			let mut min_equity = T::Decimal::zero();
 			let mut equity: T::Decimal = margin.into_decimal()?;
 			for position in positions.iter() {
-				if let Some(direction) = Self::position_direction(position) {
+				if let Some(direction) = position.direction() {
 					// Should always succeed
 					let market = Markets::<T>::get(&position.market_id)
 						.ok_or(Error::<T>::MarketIdNotFound)?;
@@ -1094,16 +1013,6 @@ pub mod pallet {
 					(position, index)
 				},
 			})
-		}
-
-		fn position_direction(position: &Position<T>) -> Option<Direction> {
-			if position.base_asset_amount.is_zero() {
-				None
-			} else if position.base_asset_amount.is_positive() {
-				Some(Direction::Long)
-			} else {
-				Some(Direction::Short)
-			}
 		}
 
 		fn base_asset_value(
